fix(family): cast xp to bigint — latent panic exposed once board became reachable
The family board panicked at r.get::<i64>("xp") because
COALESCE(int,0)+COALESCE(int,0) yields int4, not int8. It never
surfaced before because the endpoint was unreachable (missing Next
rewrite → 404, slice 18); the rewrite fix made it reachable and it
502'd on every call. Cast the sum to ::bigint so the i64 read matches.

+++ memora-api/src/handlers/family.rs
@@ -91,7 +91,7 @@ pub async fn get_board(
         "SELECT u.id,
                 COALESCE(NULLIF(TRIM(CONCAT(p.first_name, ' ', COALESCE(p.last_name, ''))), ''),
                          split_part(u.email, '@', 1)) AS name,
-                COALESCE(x.xp, 0) + COALESCE(a.xp, 0) AS xp,
+                (COALESCE(x.xp, 0) + COALESCE(a.xp, 0))::bigint AS xp,
                 (SELECT COUNT(*) FROM course_review_logs l WHERE l.user_id = u.id) AS total_reviews,
                 (SELECT COUNT(*) FROM course_review_logs l WHERE l.user_id = u.id
                    AND (l.review_time + ($1 || ' minutes')::interval)::date
